Better OS and feature guards for P3A.
