+++ browser/brave_browser_main_extra_parts.cc
@@ -6,12 +6,16 @@
 #include "brave/browser/brave_browser_main_extra_parts.h"
 
 #include "brave/browser/brave_browser_process_impl.h"
-#include "brave/browser/importer/brave_importer_p3a.h"
 #include "brave/components/brave_shields/browser/brave_shields_p3a.h"
+#include "brave/components/p3a/buildflags.h"
 #include "brave/components/p3a/brave_p3a_service.h"
+#include "third_party/widevine/cdm/buildflags.h"
+
+#if !defined(OS_ANDROID)
+#include "brave/browser/importer/brave_importer_p3a.h"
 #include "brave/components/p3a/p3a_core_metrics.h"
 #include "chrome/browser/first_run/first_run.h"
-#include "third_party/widevine/cdm/buildflags.h"
+#endif  // !defined(OS_ANDROID)
 
 #if BUILDFLAG(BUNDLE_WIDEVINE_CDM)
 #include "brave/browser/widevine/brave_widevine_bundle_manager.h"
@@ -22,16 +26,16 @@ namespace {
 // Records default values for some histograms because we want these stats to be
 // uploaded anyways. Corresponding components will write new values according
 // to their usage scenarios.
-#if !defined(OS_ANDROID)
 void RecordInitialP3AValues() {
+#if !defined(OS_ANDROID)
   if (first_run::IsChromeFirstRun()) {
     RecordImporterP3A(importer::ImporterType::TYPE_UNKNOWN);
   }
+#endif  // !defined(OS_ANDROID)
+
   brave_shields::MaybeRecordShieldsUsageP3A(brave_shields::kNeverClicked,
                                             g_browser_process->local_state());
 }
-#endif  // !defined(OS_ANDROID)
-
 
 }  // namespace
 
@@ -58,13 +62,16 @@ void BraveBrowserMainExtraParts::PreMainMessageLoopRun() {
   g_brave_browser_process->brave_widevine_bundle_manager()->StartupCheck();
 #endif
   // Disabled on mobile platforms, see for instance issues/6176
-#if !defined(OS_ANDROID)
+#if BUILDFLAG(BRAVE_P3A_ENABLED)
   // TODO(iefremov): Maybe find a better place for this initialization.
   g_brave_browser_process->brave_p3a_service()->Init();
+#endif  // BUILDFLAG(BRAVE_P3A_ENABLED)
 
   RecordInitialP3AValues();
 
+  // The code below is not supported on android.
+#if !defined(OS_ANDROID)
   brave::BraveWindowTracker::CreateInstance(g_browser_process->local_state());
   brave::BraveUptimeTracker::CreateInstance(g_browser_process->local_state());
-#endif
+#endif  // !defined(OS_ANDROID)
 }
