+++ browser/brave_browser_process_impl.cc
@@ -24,8 +24,9 @@
 #include "brave/components/brave_shields/browser/https_everywhere_service.h"
 #include "brave/components/brave_shields/browser/referrer_whitelist_service.h"
 #include "brave/components/brave_shields/browser/tracking_protection_service.h"
+#include "brave/components/p3a/buildflags.h"
+#include "brave/components/p3a/brave_histogram_rewrite.h"
 #include "brave/components/p3a/brave_p3a_service.h"
-#include "brave/components/p3a/p3a_core_metrics.h"
 #include "chrome/common/chrome_paths.h"
 #include "components/component_updater/component_updater_service.h"
 #include "components/component_updater/timer_update_scheduler.h"
@@ -83,12 +84,12 @@ BraveBrowserProcessImpl::BraveBrowserProcessImpl(StartupData* startup_data)
                      },
                      base::Unretained(brave_stats_updater())));
   // Disabled on mobile platforms, see for instance issues/6176
-#if !defined(OS_ANDROID)
+#if BUILDFLAG(BRAVE_P3A_ENABLED)
   // Create P3A Service early to catch more histograms. The full initialization
   // should be started once browser process impl is ready.
   brave_p3a_service();
   brave::SetupHistogramsBraveization();
-#endif  // !defined(OS_ANDROID)
+#endif  // BUILDFLAG(BRAVE_P3A_ENABLED)
 }
 
 brave_component_updater::BraveComponent::Delegate*
