+++ browser/brave_local_state_prefs.cc
@@ -12,10 +12,10 @@
 #include "brave/components/brave_referrals/buildflags/buildflags.h"
 #include "brave/components/brave_shields/browser/ad_block_service.h"
 #include "brave/components/brave_shields/browser/brave_shields_p3a.h"
+#include "brave/components/p3a/buildflags.h"
 #include "brave/components/p3a/brave_p3a_service.h"
-#include "brave/components/p3a/p3a_core_metrics.h"
-#include "chrome/browser/first_run/first_run.h"
 #include "chrome/common/pref_names.h"
+#include "chrome/browser/first_run/first_run.h"
 #include "components/metrics/metrics_pref_names.h"
 #include "components/prefs/pref_registry_simple.h"
 
@@ -26,6 +26,10 @@
 #if BUILDFLAG(ENABLE_TOR)
 #include "brave/browser/tor/tor_profile_service.h"
 #endif
+
+#if !defined(OS_ANDROID)
+#include "brave/components/p3a/p3a_core_metrics.h"
+#endif  // !defined(OS_ANDROID)
 
 namespace brave {
 
@@ -49,7 +53,11 @@ void RegisterLocalStatePrefs(PrefRegistrySimple* registry) {
   registry->SetDefaultPrefValue(
       metrics::prefs::kMetricsReportingEnabled,
       base::Value(GetDefaultPrefValueForMetricsReporting()));
+
+#if BUILDFLAG(BRAVE_P3A_ENABLED)
   brave::BraveP3AService::RegisterPrefs(registry);
+#endif  // BUILDFLAG(BRAVE_P3A_ENABLED)
+
   brave_shields::RegisterShieldsP3APrefs(registry);
 #if !defined(OS_ANDROID)
   BraveWindowTracker::RegisterPrefs(registry);
